requirements.txt
python-telegram-bot==20.8
qrcode
pillow
opencv-python
numpy
python-dotenv
httpx
pytz
